Number of rounds adjustable; relief adjustable; optimize
The main optimization makes use of the fact that the tests are all
division by small-ish prime factors.  If we take the least common
multiple of all the monkeys' test factors, we can always take the
modulus of a worry score and the LCM to produce a smaller worry score
that has all the same properties w/ respect to this puzzle.