Solved for 2022 day 17, part 2.
Solved for 2022 day 17, part 2.
I scribbled some notes to figure this one out.
It was pretty obvious that there was going to be some kind of cycle that
could be used to mathematically calculate the answer, as 1 trillion
loops was going to take forever.

However, the specifics of the cycle were not obvious.
My approach was to search for a pattern on a number of rocks that was
small relative to the actual number of rocks required.  I then searched
for a cycle and mapped the highest points to rock numbers.  When I found
the cycle, I matched up the rock numbers.  I also had to create a table
of rock cycle offsets so I could figure out how many trailing lines
there would be if the targt number of rocks didn't fall on an even cycle
boundary.