This works for p1.
This works for p1.

My optimization to always build a geode bot when you can was invalid
for some reason.

I got most of me key insights from studying Hyper-Nuetrino's code at
https://github.com/hyper-neutrino/advent-of-code.git.

The main thing I learned was that instead of adusting the time by a
single minute on each recursion, you can actually calculate the state
and time you'd be at by waiting to build a robot.  This eliminates a ton
of states that are not needed to solve the problem.